+++ R/index_max_acceleration.R
@@ -29,8 +29,6 @@
 #'
 
 index_max_acceleration <- function(x_vector, y_vector, absolute = FALSE) {
-  # check inputs
-  # stopifnot(is_xy_v(x_vector, y_vector), is_l_a(absolute))
 
   # distances
   distances <-