+++ R/max_ad.R
@@ -73,15 +73,6 @@ max_ad <- function(x_vector,
     y_end <- y_vector[length(x_vector)]
   }
 
-  # check inputs
-  # stopifnot(
-  #   is_xy_v(x_vector, y_vector),
-  #   is_n_a(x_start),
-  #   is_n_a(y_start),
-  #   is_n_a(x_end),
-  #   is_n_a(y_end)
-  # )
-
   # shift data
   x_shift <- x_vector - x_start
   y_shift <- y_vector - y_start
@@ -91,7 +82,6 @@ max_ad <- function(x_vector,
   m_sin <- sin(-angle)
   m_cos <- cos(-angle)
 
-  # x_rot <- (x_shift*m_cos) - (y_shift*m_sin)         # not needed
   y_rot <- (x_shift * m_sin) + (y_shift * m_cos)
 
   # find the index of maximum deviation from ideal trajectory