+++ R/point_crosses.R
@@ -27,11 +27,9 @@
 #'
 
 point_crosses <- function(numeric_vector, relevant_point = 0) {
-  # input check
-  # stopifnot(is_n_v(numeric_vector), is_n_a(relevant_point))
 
-  if (length(numeric_vector) < 3 |
-    length(numeric_vector[numeric_vector != relevant_point]) < 3) {
+  if (length(numeric_vector) < 3 ||
+        length(numeric_vector[numeric_vector != relevant_point]) < 3) {
     return(NA)
   }