+++ NEWS.md
@@ -1,5 +1,6 @@
 # mousetRajectory 0.1.0
 
+* 'time_circle_entered' and 'time_circle_left' arguments have been converted to snake case ('xMid' -> 'x_mid' and 'yMid' -> 'y_mid')
 * wuelib has been re-named into mousetRajectory! Yeah!